+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2012-2022 by the GalSim developers team on GitHub
+Copyright (c) 2012-2024 by the GalSim developers team on GitHub
 https://github.com/GalSim-developers
 
 Redistribution and use in source and binary forms, with or without
