+++ jax_galsim/moffat.py
@@ -4,8 +4,7 @@ import jax.numpy as jnp
 from jax.tree_util import Partial as partial
 from jax.tree_util import register_pytree_node_class
 
-from jax_galsim.bessel import kv
-from jax_galsim.core.bessel import j0
+from jax_galsim.bessel import j0, kv
 from jax_galsim.core.draw import draw_by_kValue, draw_by_xValue
 from jax_galsim.core.integrate import ClenshawCurtisQuad, quad_integral
 from jax_galsim.core.utils import bisect_for_root, ensure_hashable, implements
