+++ jax_galsim/random.py
@@ -700,248 +700,6 @@ class Chi2Deviate(BaseDeviate):
         return "galsim.Chi2Deviate(n=%r)" % (ensure_hashable(self.n),)
 
 
-# class DistDeviate(BaseDeviate):
-#     """A class to draw random numbers from a user-defined probability distribution.
-
-#     DistDeviate is a `BaseDeviate` class that can be used to draw from an arbitrary probability
-#     distribution.  The probability distribution passed to DistDeviate can be given one of three
-#     ways: as the name of a file containing a 2d ASCII array of x and P(x), as a `LookupTable`
-#     mapping x to P(x), or as a callable function.
-
-#     Once given a probability, DistDeviate creates a table of the cumulative probability and draws
-#     from it using a `UniformDeviate`.  The precision of its outputs can be controlled with the
-#     keyword ``npoints``, which sets the number of points DistDeviate creates for its internal table
-#     of CDF(x).  To prevent errors due to non-monotonicity, the interpolant for this internal table
-#     is always linear.
-
-#     Two keywords, ``x_min`` and ``x_max``, define the support of the function.  They must be passed
-#     if a callable function is given to DistDeviate, unless the function is a `LookupTable`, which
-#     has its own defined endpoints.  If a filename or `LookupTable` is passed to DistDeviate, the
-#     use of ``x_min`` or ``x_max`` will result in an error.
-
-#     If given a table in a file, DistDeviate will construct an interpolated `LookupTable` to obtain
-#     more finely gridded probabilities for generating the cumulative probability table.  The default
-#     ``interpolant`` is linear, but any interpolant understood by `LookupTable` may be used.  We
-#     caution against the use of splines because they can cause non-monotonic behavior.  Passing the
-#     ``interpolant`` keyword next to anything but a table in a file will result in an error.
-
-#     **Examples**:
-
-#     Some sample initialization calls::
-
-#         >>> d = galsim.DistDeviate(function=f, x_min=x_min, x_max=x_max)
-
-#     Initializes d to be a DistDeviate instance with a distribution given by the callable function
-#     ``f(x)`` from ``x=x_min`` to ``x=x_max`` and seeds the PRNG using current time::
-
-#         >>> d = galsim.DistDeviate(1062533, function=file_name, interpolant='floor')
-
-#     Initializes d to be a DistDeviate instance with a distribution given by the data in file
-#     ``file_name``, which must be a 2-column ASCII table, and seeds the PRNG using the integer
-#     seed 1062533. It generates probabilities from ``file_name`` using the interpolant 'floor'::
-
-#         >>> d = galsim.DistDeviate(rng, function=galsim.LookupTable(x,p))
-
-#     Initializes d to be a DistDeviate instance with a distribution given by P(x), defined as two
-#     arrays ``x`` and ``p`` which are used to make a callable `LookupTable`, and links the
-#     DistDeviate PRNG to the already-existing random number generator ``rng``.
-
-#     Successive calls to ``d()`` generate pseudo-random values with the given probability
-#     distribution::
-
-#         >>> d = galsim.DistDeviate(31415926, function=lambda x: 1-abs(x), x_min=-1, x_max=1)
-#         >>> d()
-#         -0.4151921102709466
-#         >>> d()
-#         -0.00909781188974034
-
-#     Parameters:
-#         seed:           Something that can seed a `BaseDeviate`: an integer seed or another
-#                         `BaseDeviate`.  Using 0 means to generate a seed from the system.
-#                         [default: None]
-#         function:       A callable function giving a probability distribution or the name of a
-#                         file containing a probability distribution as a 2-column ASCII table.
-#                         [required]
-#         x_min:          The minimum desired return value (required for non-`LookupTable`
-#                         callable functions; will raise an error if not passed in that case, or if
-#                         passed in any other case) [default: None]
-#         x_max:          The maximum desired return value (required for non-`LookupTable`
-#                         callable functions; will raise an error if not passed in that case, or if
-#                         passed in any other case) [default: None]
-#         interpolant:    Type of interpolation used for interpolating a file (causes an error if
-#                         passed alongside a callable function).  Options are given in the
-#                         documentation for `LookupTable`. [default: 'linear']
-#         npoints:        Number of points DistDeviate should create for its internal interpolation
-#                         tables. [default: 256, unless the function is a non-log `LookupTable`, in
-#                         which case it uses the table's x values]
-#     """
-#     def __init__(self, seed=None, function=None, x_min=None,
-#                  x_max=None, interpolant=None, npoints=None):
-#         from .table import LookupTable
-#         from . import utilities
-#         from . import integ
-
-#         # Set up the PRNG
-#         self._rng_type = _galsim.UniformDeviateImpl
-#         self._rng_args = ()
-#         self.reset(seed)
-
-#         # Basic input checking and setups
-#         if function is None:
-#             raise TypeError('You must pass a function to DistDeviate!')
-
-#         self._interpolant = interpolant
-#         self._npoints = npoints
-#         self._xmin = x_min
-#         self._xmax = x_max
-
-#         # Figure out if a string is a filename or something we should be using in an eval call
-#         if isinstance(function, str):
-#             self._function = function # Save the inputs to be used in repr
-#             import os.path
-#             if os.path.isfile(function):
-#                 if interpolant is None:
-#                     interpolant='linear'
-#                 if x_min or x_max:
-#                     raise GalSimIncompatibleValuesError(
-#                         "Cannot pass x_min or x_max with a filename argument",
-#                         function=function, x_min=x_min, x_max=x_max)
-#                 function = LookupTable.from_file(function, interpolant=interpolant)
-#                 x_min = function.x_min
-#                 x_max = function.x_max
-#             else:
-#                 try:
-#                     function = utilities.math_eval('lambda x : ' + function)
-#                     if x_min is not None: # is not None in case x_min=0.
-#                         function(x_min)
-#                     else:
-#                         # Somebody would be silly to pass a string for evaluation without x_min,
-#                         # but we'd like to throw reasonable errors in that case anyway
-#                         function(0.6) # A value unlikely to be a singular point of a function
-#                 except Exception as e:
-#                     raise GalSimValueError(
-#                         "String function must either be a valid filename or something that "
-#                         "can eval to a function of x.\n"
-#                         "Caught error: {0}".format(e), self._function)
-#         else:
-#             # Check that the function is actually a function
-#             if not hasattr(function, '__call__'):
-#                 raise TypeError('function must be a callable function or a string')
-#             if interpolant:
-#                 raise GalSimIncompatibleValuesError(
-#                     "Cannot provide an interpolant with a callable function argument",
-#                     interpolant=interpolant, function=function)
-#             if isinstance(function, LookupTable):
-#                 if x_min or x_max:
-#                     raise GalSimIncompatibleValuesError(
-#                         "Cannot provide x_min or x_max with a LookupTable function",
-#                         function=function, x_min=x_min, x_max=x_max)
-#                 x_min = function.x_min
-#                 x_max = function.x_max
-#             else:
-#                 if x_min is None or x_max is None:
-#                     raise GalSimIncompatibleValuesError(
-#                         "Must provide x_min and x_max when function argument is a regular "
-#                         "python callable function",
-#                         function=function, x_min=x_min, x_max=x_max)
-
-#             self._function = function # Save the inputs to be used in repr
-
-#         # Compute the probability distribution function, pdf(x)
-#         if (npoints is None and isinstance(function, LookupTable) and
-#             not function.x_log and not function.f_log):
-#             xarray = np.array(function.x, dtype=float)
-#             pdf = np.array(function.f, dtype=float)
-#             # Set up pdf, so cumsum basically does a cumulative trapz integral
-#             # On Python 3.4, doing pdf[1:] += pdf[:-1] the last value gets messed up.
-#             # Writing it this way works.  (Maybe slightly slower though, so if we stop
-#             # supporting python 3.4, consider switching to the += version.)
-#             pdf[1:] = pdf[1:] + pdf[:-1]
-#             pdf[1:] *= np.diff(xarray)
-#             pdf[0] = 0.
-#         else:
-#             if npoints is None: npoints = 256
-#             xarray = x_min+(1.*x_max-x_min)/(npoints-1)*np.array(range(npoints),float)
-#             # Integrate over the range of x in case the function is doing something weird here.
-#             pdf = [0.] + [integ.int1d(function, xarray[i], xarray[i+1])
-#                           for i in range(npoints - 1)]
-#             pdf = np.array(pdf)
-
-#         # Check that the probability is nonnegative
-#         if not np.all(pdf >= 0.):
-#             raise GalSimValueError('Negative probability found in DistDeviate.',function)
-
-#         # Compute the cumulative distribution function = int(pdf(x),x)
-#         cdf = np.cumsum(pdf)
-
-#         # Quietly renormalize the probability if it wasn't already normalized
-#         totalprobability = cdf[-1]
-#         cdf /= totalprobability
-
-#         self._inverse_cdf = LookupTable(cdf, xarray, interpolant='linear')
-#         self.x_min = x_min
-#         self.x_max = x_max
-
-#     def val(self, p):
-#         r"""
-#         Return the value :math:`x` of the input function to `DistDeviate` such that ``p`` =
-#         :math:`F(x)`, where :math:`F` is the cumulattive probability distribution function:
-
-#         .. math::
-
-#             F(x) = \int_{-\infty}^x \mathrm{pdf}(t) dt
-
-#         This function is typically called by `__call__`, which generates a random p
-#         between 0 and 1 and calls ``self.val(p)``.
-
-#         Parameters:
-#             p:      The desired cumulative probabilty p.
-
-#         Returns:
-#             the corresponding x such that :math:`p = F(x)`.
-#         """
-#         if p<0 or p>1:
-#             raise GalSimRangeError('Invalid cumulative probability for DistDeviate', p, 0., 1.)
-#         return self._inverse_cdf(p)
-
-#     def __call__(self):
-#         """Draw a new random number from the distribution.
-#         """
-#         return self._inverse_cdf(self._rng.generate1())
-
-#     def generate(self, array):
-#         """Generate many pseudo-random values, filling in the values of a numpy array.
-#         """
-#         p = np.empty_like(array)
-#         BaseDeviate.generate(self, p)  # Fill with unform deviate values
-#         np.copyto(array, self._inverse_cdf(p)) # Convert from p -> x
-
-#     def add_generate(self, array):
-#         """Generate many pseudo-random values, adding them to the values of a numpy array.
-#         """
-#         p = np.empty_like(array)
-#         BaseDeviate.generate(self, p)
-#         array += self._inverse_cdf(p)
-
-#     def __repr__(self):
-#         return ('galsim.DistDeviate(seed=%r, function=%r, x_min=%r, x_max=%r, interpolant=%r, '
-#                 'npoints=%r)')%(self._seed_repr(), self._function, self._xmin, self._xmax,
-#                                 self._interpolant, self._npoints)
-#     def __str__(self):
-#         return 'galsim.DistDeviate(function="%s", x_min=%s, x_max=%s, interpolant=%s, npoints=%s)'%(
-#                 self._function, self._xmin, self._xmax, self._interpolant, self._npoints)
-
-#     def __eq__(self, other):
-#         return (self is other or
-#                 (isinstance(other, DistDeviate) and
-#                  self.serialize() == other.serialize() and
-#                  self._function == other._function and
-#                  self._xmin == other._xmin and
-#                  self._xmax == other._xmax and
-#                  self._interpolant == other._interpolant and
-#                  self._npoints == other._npoints))
-
-
 @implements(
     _galsim.random.permute,
     lax_description="The JAX implementation of this function cannot operate in-place and so returns a new list of arrays.",
