+++ jax_galsim/sum.py
@@ -188,6 +188,7 @@ class Sum(GSObject):
     def _flux_per_photon(self):
         return self._calculate_flux_per_photon()
 
+    @implements(_galsim.Sum._shoot)
     def _shoot(self, photons, rng):
         tot_flux = self.positive_flux + self.negative_flux
         fluxes = jnp.array(
